Split up jinja2 extensions and simplify them
